Extract csv importer to separate class
