+++ test/fixtures/files/reference_library_documents_test.csv
@@ -1,2 +1,3 @@
-﻿Name & Organization,Attachments,Document Title,Description,Technical Area,Indicator,Document Type,Author,Publication Date,WHO-Approved Doc: Check here if this document peer reviewed or from the WHO or CDC,Status,Relevant Pages,Lead Approval Req: Check here if this document requires approval from another lead agency,Thumbnail Image
+Name & Organization,Attachments,Document Title,Description,Technical Area,Indicator,Document Type,Author,Publication Date,WHO-Approved Doc: Check here if this document peer reviewed or from the WHO or CDC,Status,Relevant Pages,Lead Approval Req: Check here if this document requires approval from another lead agency,Thumbnail Image
 Michael Mahar,Hazim Ethiopia AMR Surveillance.pdf (https://dl.airtable.com/.attachments/7db0812500c513bb055fe551030f0d84/b9d0e9bb/HazimEthiopiaAMRSurveillance.pdf),Establishment of a Sentinel Laboratory-Based Antimicrobial Resistance Surveillance Network in Ethiopia,Description of establishment of an AMR sentinel surveillance system in Ethiopia.,Antimicrobial Resistance,3.2 Initiate AMR surveillance at pilot or representative regional and referral hospitals.,Guideline,Hazim et al.,2018,checked,,All,,Hazim Ethiopia AMR Surveillance.png (https://dl.airtable.com/.attachments/5eaacdc47b1401fa8ff87cf373cb6474/ebb914f1/HazimEthiopiaAMRSurveillance.png)
+Britney McMurren,AMS Toolkit_Eng.pdf (https://dl.airtable.com/.attachments/16400640c2b8abf305043934722d0a6f/239bacdd/AMSToolkit_Eng.pdf),Antimicrobial Stewardship Programmes in Health-care Facilities in Low- and Middle-Income Countries: A WHO Practical Toolkit,The specific aim of the toolkit is to enable antimicrobial stewardship in health-care facilities in low- and middle-income countries.,Antimicrobial Resistance,"3.4 Expand antimicrobial stewardship activities to all healthcare facilities.,""3.4 Monitor and evaluate stewardship programmes conducted, including the analysis of antimicrobial use data."",3.4 Implement stewardship practices at designated healthcare facilities.",Tool,World Health Organization ,2019,checked,,All,,AMS Toolkit_Eng.png (https://dl.airtable.com/.attachments/e7498e320b5151eb4c32b9b86ac0cfec/b1c0396a/AMSToolkit_Eng.png)
